created restart script for cronjob

+++ restart.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
+RUNNING=`docker-compose top`
+if [[ "$RUNNING" != "" ]]; then
 docker-compose down
+fi
 docker image prune -f
-git pull -f
 docker-compose up -d --build
